test(reader): split quick selector coverage
Why:
- Continue moving pure BibleUI reader behavior tests out of the app-host XCTest bundle so they run in the package test lane.
- Preserve Android quick document selector and current-document switch contracts while reducing simulator-bound unit coverage.

What Changed:
- Added package-level quick module selector presentation/action coverage.
- Added package-level controller document-switch side-effect coverage with temporary SWORD fixtures.
- Removed the migrated tests from AndBibleTests+AppAndReader while leaving app-host source guards for a later slice.
- Updated the migration plan with the completed quick-selector slice and remaining AppAndReader scope.

Validation:
- python3 scripts/check_repo_standards.py docblocks --all-files
- python3 scripts/check_repo_standards.py source-guards
- git diff --check
- GitNexus detect_changes scope=staged
- xcodebuild test -project AndBible.xcodeproj -scheme BibleUITests -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-quick-selector-bibleui-dd CODE_SIGNING_ALLOWED=NO -only-testing:BibleUITests/BibleReaderQuickModuleSelectorTests -only-testing:BibleUITests/BibleReaderDocumentSwitchControllerTests
- xcodebuild test -quiet -project AndBible.xcodeproj -scheme AndBible -destination 'platform=iOS Simulator,name=iPhone 17' -derivedDataPath /tmp/andbible-quick-selector-app-dd CODE_SIGNING_ALLOWED=NO -only-testing:AndBibleTests/AndBibleTests/testBibleToolbarMenuRoutesThroughAnchoredQuickSelectorInsteadOfSheet -only-testing:AndBibleTests/AndBibleTests/testApplicationDelegateSceneConfigurationUsesWindowSceneDelegate

Impact:
- Test-only migration; GitNexus reported no mapped production-flow changes.
- Android quick selector parity coverage remains in BibleUI package tests.
- App-host coverage is narrowed without changing production behavior.
