+++ AndBibleTests/AndBibleTests+AppAndReader.swift
@@ -17,499 +17,6 @@ import struct SwiftUI.Color
 
 extension AndBibleTests {
     /**
-     Protects Android `MainBibleActivity.menuForDocs` parity for the Bible toolbar quick menu.
-
-     Android sorts quick-menu entries by language code and then book abbreviation, renders labels as
-     abbreviation plus language code in parentheses, and disables the current document instead of
-     re-selecting it. This test uses the pure presentation contract so future UI refactors cannot
-     accidentally restore the old iOS full-sheet semantics or sort by localized description.
-     */
-    func testBibleQuickModuleSelectorRowsMirrorAndroidOrderingLabelsAndDisabledCurrentDocument() {
-        let modules = [
-            ModuleInfo(name: "WEB", description: "World English Bible", category: .bible, language: "en"),
-            ModuleInfo(name: "FinRK", description: "Finnish Revised Version", category: .bible, language: "fi"),
-            ModuleInfo(name: "AB", description: "Another Bible", category: .bible, language: "en")
-        ]
-
-        let rows = BibleReaderQuickModuleSelectorPresentation.rows(
-            for: modules,
-            activeModuleName: "WEB"
-        )
-
-        XCTAssertEqual(rows.map(\.module.name), ["AB", "WEB", "FinRK"])
-        XCTAssertEqual(rows.map(\.title), ["AB (en)", "WEB (en)", "FinRK (fi)"])
-        XCTAssertEqual(rows.map(\.isEnabled), [true, false, true])
-    }
-
-    /**
-     Protects the quick-selector row equality contract used by presentation tests.
-
-     Row equality should cover only visible and behavior-significant fields: selected module name,
-     rendered title, and enabled state. Metadata such as description, category, or language is
-     normalized into the title before rendering and should not make tests fail when behavior is
-     unchanged.
-     */
-    func testBibleQuickModuleSelectorRowEqualityIgnoresNonVisibleModuleMetadata() {
-        let lhs = BibleReaderQuickModuleSelectorPresentation.Row(
-            module: ModuleInfo(
-                name: "KJV",
-                description: "King James Version",
-                category: .bible,
-                language: "en"
-            ),
-            title: "KJV (en)",
-            isEnabled: true
-        )
-        let rhs = BibleReaderQuickModuleSelectorPresentation.Row(
-            module: ModuleInfo(
-                name: "KJV",
-                description: "Different catalog description",
-                category: .commentary,
-                language: "fi"
-            ),
-            title: "KJV (en)",
-            isEnabled: true
-        )
-
-        XCTAssertEqual(lhs, rhs)
-    }
-
-    /**
-     Protects long quick-selector lists from becoming unscrollable off-screen stacks.
-
-     Users can install dozens of Bible modules. Android renders those entries through a popup menu
-     that can scroll; the SwiftUI parity renderer must likewise be backed by a scroll container so
-     all available modules can be reached without falling back to the full iOS sheet.
-     */
-    func testBibleQuickModuleSelectorUsesScrollContainerForLongInstalledModuleLists() throws {
-        let rows = (0..<60).map { index in
-            BibleReaderQuickModuleSelectorPresentation.Row(
-                module: ModuleInfo(
-                    name: String(format: "MOD%02d", index),
-                    description: "Module \(index)",
-                    category: .bible,
-                    language: "en"
-                ),
-                title: String(format: "MOD%02d (en)", index),
-                isEnabled: true
-            )
-        }
-        let view = BibleReaderQuickModuleSelector(
-            rows: rows,
-            colorScheme: .light,
-            onSelect: { _ in }
-        )
-
-        XCTAssertTrue(String(describing: type(of: view.body)).contains("ScrollView"))
-        let selectorSource = try bibleUISource(named: "BibleReaderQuickModuleSelector.swift")
-        XCTAssertTrue(selectorSource.contains("LazyVStack(alignment: .leading, spacing: 0)"))
-        XCTAssertTrue(selectorSource.contains("Button {"))
-        XCTAssertTrue(selectorSource.contains(".buttonStyle(.plain)"))
-        XCTAssertTrue(selectorSource.contains(".disabled(!row.isEnabled)"))
-        XCTAssertFalse(selectorSource.contains("            VStack(alignment: .leading, spacing: 0)"))
-    }
-
-    /**
-     Protects Android's exactly-two-document shortcut in `menuForDocs`.
-
-     When only two Bible modules are available, Android switches directly to the other document and
-     does not show a popup. The iOS toolbar action must keep that shortcut while replacing only the
-     three-or-more path with the compact quick selector.
-     */
-    func testBibleQuickModuleSelectorActionMirrorsAndroidTwoDocumentShortcut() {
-        let modules = [
-            ModuleInfo(name: "KJV", description: "King James Version", category: .bible, language: "en"),
-            ModuleInfo(name: "WEB", description: "World English Bible", category: .bible, language: "en")
-        ]
-
-        let action = BibleReaderQuickModuleSelectorPresentation.action(
-            for: modules,
-            activeModuleName: "KJV"
-        )
-
-        guard case .switchDirectly(let row) = action else {
-            XCTFail("Expected Android's exactly-two-module shortcut to select the alternate Bible.")
-            return
-        }
-        XCTAssertEqual(row.module.name, "WEB")
-        XCTAssertEqual(row.module.category, .bible)
-    }
-
-    /**
-     Protects Android's popup threshold in `menuForDocs`.
-
-     Three or more Bible modules must show the compact anchored quick selector, not the full
-     document picker sheet. The sorted rows are part of the action payload so the UI layer cannot
-     accidentally diverge from Android ordering while still showing a popup.
-     */
-    func testBibleQuickModuleSelectorActionShowsPopupForMoreThanTwoModules() {
-        let modules = [
-            ModuleInfo(name: "WEB", description: "World English Bible", category: .bible, language: "en"),
-            ModuleInfo(name: "FinRK", description: "Finnish Revised Version", category: .bible, language: "fi"),
-            ModuleInfo(name: "AB", description: "Another Bible", category: .bible, language: "en")
-        ]
-
-        let action = BibleReaderQuickModuleSelectorPresentation.action(
-            for: modules,
-            activeModuleName: "WEB"
-        )
-
-        XCTAssertEqual(
-            action,
-            .showPopup([
-                BibleReaderQuickModuleSelectorPresentation.Row(
-                    module: modules[2],
-                    title: "AB (en)",
-                    isEnabled: true
-                ),
-                BibleReaderQuickModuleSelectorPresentation.Row(
-                    module: modules[0],
-                    title: "WEB (en)",
-                    isEnabled: false
-                ),
-                BibleReaderQuickModuleSelectorPresentation.Row(
-                    module: modules[1],
-                    title: "FinRK (fi)",
-                    isEnabled: true
-                )
-            ])
-        )
-    }
-
-    /**
-     Protects Android's non-two-document popup rule for single available Bible modules.
-
-     Android only special-cases exactly two documents. With one available Bible it still shows the
-     popup, and if the current document is not that Bible then the row remains enabled so the toolbar
-     can switch from commentary or another category back to Bible mode.
-     */
-    func testBibleQuickModuleSelectorActionShowsPopupForSingleModuleWhenBibleIsNotCurrentDocument() {
-        let modules = [
-            ModuleInfo(name: "KJV", description: "King James Version", category: .bible, language: "en")
-        ]
-
-        let action = BibleReaderQuickModuleSelectorPresentation.action(
-            for: modules,
-            activeModuleName: nil
-        )
-
-        XCTAssertEqual(
-            action,
-            .showPopup([
-                BibleReaderQuickModuleSelectorPresentation.Row(
-                    module: modules[0],
-                    title: "KJV (en)",
-                    isEnabled: true
-                )
-            ])
-        )
-    }
-
-    /**
-     Protects the quick selector's selected-row side effect contract.
-
-     Android's popup selection calls the same current-document switch path used elsewhere in the
-     reader. iOS should likewise route selected quick-menu modules through
-     `BibleReaderController.switchBibleDocument(to:)` and persist the chosen Bible document plus
-     Bible category on the pane's `PageManager`, rather than maintaining separate quick-selector
-     state.
-     */
-    @MainActor
-    func testBibleQuickModuleSelectorSelectionUsesControllerSwitchPathAndPersistsPaneDocument() throws {
-        let (bridge, _) = makeRecordingBridge()
-        let modulePath = try makeTemporaryBundledSwordPath()
-        try seedBibleAliasModule(
-            named: "WEB",
-            description: "World English Bible",
-            in: modulePath
-        )
-        let manager = try XCTUnwrap(SwordManager(modulePath: modulePath))
-        let controller = BibleReaderController(bridge: bridge, swordManagerOverride: manager)
-        let window = Window()
-        let pageManager = PageManager(id: window.id)
-        window.pageManager = pageManager
-        controller.activeWindow = window
-
-        let action = BibleReaderQuickModuleSelectorPresentation.action(
-            for: controller.installedBibleModules,
-            activeModuleName: controller.activeModuleName
-        )
-        guard case .switchDirectly(let row) = action else {
-            XCTFail("Expected Android's exactly-two-module shortcut to select the alternate Bible.")
-            return
-        }
-
-        controller.switchBibleDocument(to: row.module.name)
-
-        XCTAssertEqual(controller.activeModuleName, "WEB")
-        XCTAssertEqual(pageManager.bibleDocument, "WEB")
-        XCTAssertEqual(pageManager.currentCategoryName, DocumentCategory.bible.pageManagerKey)
-    }
-
-    /**
-     Protects Android `MainBibleActivity.commentaryClick` candidate and row semantics.
-
-     The Android default commentary toolbar tap calls `menuForDocs` with unlocked commentaries plus
-     general books plus dictionaries, then `menuForDocs` sorts by language code and abbreviation and
-     renders compact `initials (language)` rows. This test keeps that category mix in the shared
-     quick-selector presentation contract so iOS cannot regress to the full commentary-only chooser
-     sheet or sort by localized descriptions.
-     */
-    func testCommentaryQuickModuleSelectorRowsIncludeAndroidDocumentCategories() {
-        let commentary = ModuleInfo(
-            name: "MHC",
-            description: "Matthew Henry",
-            category: .commentary,
-            language: "en"
-        )
-        let dictionary = ModuleInfo(
-            name: "BDBT",
-            description: "Brown Driver Briggs",
-            category: .dictionary,
-            language: "en"
-        )
-        let generalBook = ModuleInfo(
-            name: "Pilgrim",
-            description: "Pilgrim's Progress",
-            category: .generalBook,
-            language: "en"
-        )
-        let finnishCommentary = ModuleInfo(
-            name: "FinComm",
-            description: "Finnish Commentary",
-            category: .commentary,
-            language: "fi"
-        )
-
-        let rows = BibleReaderQuickModuleSelectorPresentation.rows(
-            for: [generalBook, finnishCommentary, commentary, dictionary],
-            activeModuleName: "BDBT"
-        )
-
-        XCTAssertEqual(rows.map(\.module.category), [.dictionary, .commentary, .generalBook, .commentary])
-        XCTAssertEqual(rows.map(\.title), ["BDBT (en)", "MHC (en)", "Pilgrim (en)", "FinComm (fi)"])
-        XCTAssertEqual(rows.map(\.isEnabled), [false, true, true, true])
-    }
-
-    /**
-     Protects Android's atomic current-document switch behavior for commentary quick selections.
-
-     Android `menuForDocs` delegates the selected commentary `Book` to `setCurrentDocument(book)`,
-     which updates the active document and visible category together. iOS must provide the same
-     controller-level contract for the quick selector instead of switching module and category in
-     separate calls that can reload stale content or persist partial pane state.
-     */
-    @MainActor
-    func testCommentaryDocumentSwitchPersistsModuleAndCategoryTogether() throws {
-        let (bridge, _) = makeRecordingBridge()
-        let modulePath = try makeTemporaryBundledSwordPath()
-        try seedEmptyRawCommentaryModule(named: "UITestComm", in: modulePath)
-        let manager = try XCTUnwrap(SwordManager(modulePath: modulePath))
-        let controller = BibleReaderController(bridge: bridge, swordManagerOverride: manager)
-        let window = Window()
-        let pageManager = PageManager(id: window.id)
-        window.pageManager = pageManager
-        controller.activeWindow = window
-        var persistCount = 0
-        controller.onPersistState = { persistCount += 1 }
-
-        controller.switchCommentaryDocument(to: "UITestComm")
-
-        XCTAssertEqual(controller.currentCategory, .commentary)
-        XCTAssertEqual(controller.activeCommentaryModuleName, "UITestComm")
-        XCTAssertEqual(pageManager.commentaryDocument, "UITestComm")
-        XCTAssertEqual(pageManager.currentCategoryName, DocumentCategory.commentary.pageManagerKey)
-        XCTAssertEqual(persistCount, 1)
-    }
-
-    /**
-     Protects Android's atomic current-document switch behavior for dictionary quick selections.
-
-     Android routes dictionaries from the commentary quick popup through `setCurrentDocument(book)`.
-     iOS must therefore persist the selected dictionary, clear stale dictionary entry state, and
-     switch the visible category in one controller call rather than splitting module and category
-     updates across separate mutations.
-     */
-    @MainActor
-    func testDictionaryDocumentSwitchPersistsModuleCategoryAndClearsKeyTogether() throws {
-        let (bridge, _) = makeRecordingBridge()
-        let modulePath = try makeTemporaryBundledSwordPath()
-        try seedEmptyRawDictionaryModule(named: "UITestDict", in: modulePath)
-        let manager = try XCTUnwrap(SwordManager(modulePath: modulePath))
-        let controller = BibleReaderController(bridge: bridge, swordManagerOverride: manager)
-        let window = Window()
-        let pageManager = PageManager(id: window.id)
-        pageManager.dictionaryKey = "stale-key"
-        window.pageManager = pageManager
-        controller.activeWindow = window
-        var persistCount = 0
-        controller.onPersistState = { persistCount += 1 }
-
-        controller.switchDictionaryDocument(to: "UITestDict")
-
-        XCTAssertEqual(controller.currentCategory, .dictionary)
-        XCTAssertEqual(controller.activeDictionaryModuleName, "UITestDict")
-        XCTAssertNil(controller.currentDictionaryKey)
-        XCTAssertEqual(pageManager.dictionaryDocument, "UITestDict")
-        XCTAssertNil(pageManager.dictionaryKey)
-        XCTAssertEqual(pageManager.currentCategoryName, DocumentCategory.dictionary.pageManagerKey)
-        XCTAssertEqual(persistCount, 1)
-    }
-
-    /**
-     Protects Android's atomic current-document switch behavior for general-book quick selections.
-
-     Android includes general books in the commentary quick popup and applies selections through the
-     same current-document transition. The iOS controller must persist module/category and clear the
-     stale general-book key together so quick selection cannot leave mixed pane state behind.
-     */
-    @MainActor
-    func testGeneralBookDocumentSwitchPersistsModuleCategoryAndClearsKeyTogether() throws {
-        let (bridge, _) = makeRecordingBridge()
-        let modulePath = try makeTemporaryBundledSwordPath()
-        try seedEmptyRawGeneralBookModule(named: "UITestGB", in: modulePath)
-        let manager = try XCTUnwrap(SwordManager(modulePath: modulePath))
-        let controller = BibleReaderController(bridge: bridge, swordManagerOverride: manager)
-        let window = Window()
-        let pageManager = PageManager(id: window.id)
-        pageManager.generalBookKey = "stale-key"
-        window.pageManager = pageManager
-        controller.activeWindow = window
-        var persistCount = 0
-        controller.onPersistState = { persistCount += 1 }
-
-        controller.switchGeneralBookDocument(to: "UITestGB")
-
-        XCTAssertEqual(controller.currentCategory, .generalBook)
-        XCTAssertEqual(controller.activeGeneralBookModuleName, "UITestGB")
-        XCTAssertNil(controller.currentGeneralBookKey)
-        XCTAssertEqual(pageManager.generalBookDocument, "UITestGB")
-        XCTAssertNil(pageManager.generalBookKey)
-        XCTAssertEqual(pageManager.currentCategoryName, DocumentCategory.generalBook.pageManagerKey)
-        XCTAssertEqual(persistCount, 1)
-    }
-
-    /**
-     Protects Android's atomic current-document switch behavior for map selections.
-
-     Android routes maps through the same current-document transition as other chooser rows. The
-     iOS controller must therefore persist the selected map/category and clear stale map keys
-     together so map selection cannot leave mixed pane state behind.
-     */
-    @MainActor
-    func testMapDocumentSwitchPersistsModuleCategoryAndClearsKeyTogether() throws {
-        let (bridge, _) = makeRecordingBridge()
-        let modulePath = try makeTemporaryBundledSwordPath()
-        try seedEmptyRawMapModule(named: "UITestMap", in: modulePath)
-        let manager = try XCTUnwrap(SwordManager(modulePath: modulePath))
-        let controller = BibleReaderController(bridge: bridge, swordManagerOverride: manager)
-        let window = Window()
-        let pageManager = PageManager(id: window.id)
-        pageManager.mapKey = "stale-key"
-        window.pageManager = pageManager
-        controller.activeWindow = window
-        var persistCount = 0
-        controller.onPersistState = { persistCount += 1 }
-
-        controller.switchMapDocument(to: "UITestMap")
-
-        XCTAssertEqual(controller.currentCategory, .map)
-        XCTAssertEqual(controller.activeMapModuleName, "UITestMap")
-        XCTAssertNil(controller.currentMapKey)
-        XCTAssertEqual(pageManager.mapDocument, "UITestMap")
-        XCTAssertNil(pageManager.mapKey)
-        XCTAssertEqual(pageManager.currentCategoryName, DocumentCategory.map.pageManagerKey)
-        XCTAssertEqual(persistCount, 1)
-    }
-
-    /**
-     Protects Android's atomic current-document switch behavior for Bible quick selections.
-
-     Android `MainBibleActivity.menuForDocs` delegates selected Bible rows to
-     `CurrentPageManager.setCurrentDocument(book)`, which updates the active document and page
-     category as one transition. iOS must not first reload the current non-Bible category and then
-     reload the selected Bible, because that creates unnecessary WebView work and visible flicker.
-     */
-    @MainActor
-    func testBibleDocumentSwitchFromCommentaryReloadsSelectedBibleOnce() throws {
-        let (bridge, recordedScripts) = makeRecordingBridge()
-        let modulePath = try makeTemporaryBundledSwordPath()
-        try seedBibleAliasModule(
-            named: "WEB",
-            description: "World English Bible",
-            in: modulePath
-        )
-        try seedEmptyRawCommentaryModule(in: modulePath)
-        let manager = try XCTUnwrap(SwordManager(modulePath: modulePath))
-        let controller = BibleReaderController(bridge: bridge, swordManagerOverride: manager)
-        let window = Window()
-        let pageManager = PageManager(id: window.id)
-        window.pageManager = pageManager
-        controller.activeWindow = window
-        controller.switchCategory(to: .commentary)
-        controller.bridgeDidSetClientReady(bridge)
-        let baselineScriptCount = recordedScripts().count
-        var persistCount = 0
-        controller.onPersistState = { persistCount += 1 }
-
-        controller.switchBibleDocument(to: "WEB")
-        RunLoop.main.run(until: Date(timeIntervalSinceNow: 0.2))
-
-        let addDocumentScripts = recordedScripts()
-            .dropFirst(baselineScriptCount)
-            .filter { $0.contains("emit('add_documents'") }
-        XCTAssertEqual(addDocumentScripts.count, 1)
-        let payload = try XCTUnwrap(
-            bridgeEmissionPayload(from: Array(addDocumentScripts), event: "add_documents") as? [String: Any]
-        )
-        XCTAssertEqual(payload["bookCategory"] as? String, "BIBLE")
-        XCTAssertEqual(payload["bookInitials"] as? String, "WEB")
-        XCTAssertEqual(controller.currentCategory, .bible)
-        XCTAssertEqual(controller.activeModuleName, "WEB")
-        XCTAssertEqual(pageManager.bibleDocument, "WEB")
-        XCTAssertEqual(pageManager.currentCategoryName, DocumentCategory.bible.pageManagerKey)
-        XCTAssertEqual(persistCount, 1)
-    }
-
-    /**
-     Protects the controller-level Bible switch API from accepting non-Bible modules.
-
-     The quick selector and module picker currently pass Bible-filtered rows, but
-     `BibleReaderController.switchBibleDocument(to:)` is public controller API and mirrors Android's
-     current-document transition only for Bible documents. A non-Bible SWORD module must therefore
-     leave the active Bible, document category, PageManager state, and persistence callbacks
-     unchanged. A failure means an accidental non-Bible caller can corrupt pane state by forcing the
-     reader into Bible mode with a commentary/dictionary module name.
-     */
-    @MainActor
-    func testBibleDocumentSwitchRejectsNonBibleModulesWithoutStateMutation() throws {
-        let (bridge, _) = makeRecordingBridge()
-        let modulePath = try makeTemporaryBundledSwordPath()
-        try seedEmptyRawCommentaryModule(in: modulePath)
-        let manager = try XCTUnwrap(SwordManager(modulePath: modulePath))
-        let controller = BibleReaderController(bridge: bridge, swordManagerOverride: manager)
-        let window = Window()
-        let pageManager = PageManager(id: window.id)
-        window.pageManager = pageManager
-        controller.activeWindow = window
-        controller.switchCategory(to: .commentary)
-        let baselineCategory = controller.currentCategory
-        let baselineBibleModuleName = controller.activeModuleName
-        let baselineBibleDocument = pageManager.bibleDocument
-        let baselineCategoryName = pageManager.currentCategoryName
-        var persistCount = 0
-        controller.onPersistState = { persistCount += 1 }
-
-        controller.switchBibleDocument(to: "UITestComm")
-
-        XCTAssertEqual(controller.currentCategory, baselineCategory)
-        XCTAssertEqual(controller.activeModuleName, baselineBibleModuleName)
-        XCTAssertEqual(pageManager.bibleDocument, baselineBibleDocument)
-        XCTAssertEqual(pageManager.currentCategoryName, baselineCategoryName)
-        XCTAssertEqual(persistCount, 0)
-    }
-
-    /**
      Guards the reader coordinator against regressing to the iOS sheet for the quick-menu path.
 
      The coordinator state is intentionally private, so this source-level test checks the routing
